Revert moving plot implementation to plot_impl
Reverts moving the plot implementation to plot_impl. This is because
plot uses __init__ correctly to define the public API, so users can't
easily pull in implementation details.
