+++ python/src/scipp/plot/plot_1d.py
