+++ python/src/scipp/plot/plot_2d.py
