--- python/src/scipp/plot/plot_impl/__init__.py
@@ -1,2 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0-or-later
-# Copyright (c) 2020 Scipp contributors (https://github.com/scipp)
