+++ python/src/scipp/plot/plot_request.py
